+++ cdl/core/io/__init__.py
@@ -19,45 +19,69 @@ from cdl.core.model.image import ImageObj
 from cdl.core.model.signal import SignalObj
 
 
+def read_signals(filename: str) -> list[SignalObj]:
+    """Read a list of signals from a file.
+
+    Args:
+        filename: File name.
+
+    Returns:
+        List of signals.
+    """
+    return SignalIORegistry.read(filename)
+
+
 def read_signal(filename: str) -> SignalObj:
     """Read a signal from a file.
 
     Args:
-        filename (str): File name.
+        filename: File name.
 
     Returns:
-        Signal: Signal.
+        Signal.
     """
-    return SignalIORegistry.read(filename)
+    return read_signals(filename)[0]
 
 
 def write_signal(filename: str, signal: SignalObj) -> None:
     """Write a signal to a file.
 
     Args:
-        filename (str): File name.
-        signal (Signal): Signal.
+        filename: File name.
+        signal: Signal.
     """
     SignalIORegistry.write(filename, signal)
+
+
+def read_images(filename: str) -> list[ImageObj]:
+    """Read a list of images from a file.
+
+    Args:
+        filename: File name.
+
+    Returns:
+        List of images.
+    """
+    return ImageIORegistry.read(filename)
 
 
 def read_image(filename: str) -> ImageObj:
     """Read an image from a file.
 
     Args:
-        filename (str): File name.
+        filename: File name.
 
     Returns:
-        Image: Image.
+        Image.
     """
-    return ImageIORegistry.read(filename)
+    return read_images(filename)[0]
 
 
 def write_image(filename: str, image: ImageObj) -> None:
     """Write an image to a file.
 
     Args:
-        filename (str): File name.
-        image (Image): Image.
+        filename: File name.
+        image: Image.
     """
     ImageIORegistry.write(filename, image)